feat: switch from HTTPoison to Mint
Since we don't need all the pooling features that HTTPoison & Hackney
provide, we can use the architecturally simpler Mint.

Mint is a "Small and composable HTTP client", which uses a connection struct
to maintain the state. As messages come into our GenStage server, we pass
them to Mint where they're parsed into responses.

The main change is to redirection, which we need to handle ourselves now.